Implemented exception handling. It works as follows. Defined a `TaskManager` class in `utils` that can be used to create and cancel tasks. Upon instantiation a task running the `TaskManager._monitor_tasks` coroutine is started, which continually checks if any tasks stop, and whether they raised any exceptions.
It is possible to subscribe to a TaskManager via `TaskManager.subscribe`. You pass a callback function that will be called when any exceptions are found in its tasks.

All `BaseComponent`s come with a TaskManager. It will only really use it if it runs background tasks (like `CompositeComponent`s do). Its parent `Node` subscribes to the component's TaskManager with its callback `Node._handle_component_process_exception`:

```python
def _handle_component_process_exception(self, task:asyncio.Task, exception:Exception, source_trace:Tuple[Union[BaseComponent], ...]):
    try:
        raise ComponentError() from exception
    except ComponentError as e:
        if not self.states.stopping.get(): self.stop()
        self._task_manager.submit_exception(task, e, source_trace)
```

The Node will raise a `ComponentError` and submit the exception to its own TaskManager. If the Node is not the Net (i.e. if it has a parent `CompositeComponent`), then `CompositeComponent._handle_child_node_exception` will be callbacked in turn, which will raise a `NodeError`. This will eventually reach the top node, the net, in a chain of exceptions.

All `asyncio.CancelledError` must be handled. That means that if a task is cancelled, and the task does not catch `asyncio.CancelledError` the exception will bubble up to the Net, after which it would be raised (if the Net is run via `Net.execute`.

In `Net` there are also now `Net.execute` and `Net.async_execute` methods, as well as the class method `Net.execute_graph`. Nets should now always be top nodes. Nets are essentially conceived as wrappers around Node, that make them more amenable for various use cases. `Net` for example turns a running Node into something you can call like a normal function. Given this, `Net` should probably be called `BatchNet` or something instead. You could have a `CLINet` that turns a graph into a CLI command, for example.

Considering that Nets are just wrappers, I might eventually rename `Net` into something like `BatchRuntime` and not make it a subclass of `Node`.

---

The main execution code of `Net` is as follows:

```python
async def main():
    self.run()

    async def net_IO_handler():
        input_sender_task = asyncio.create_task(input_sender())
        output_receiver_task = asyncio.create_task(output_receiver())
        await asyncio.gather(input_sender_task, output_receiver_task)

    _, _ = await asyncio.wait(
        [
            asyncio.create_task(net_IO_handler()),
            asyncio.create_task(self._task_manager.wait_for_exceptions())
        ],
        return_when=asyncio.FIRST_COMPLETED
    )

    if self._task_manager.exception_occured():
        for task, e, source_trace in self._task_manager.get_exceptions():
            raise e

    _, _ = await asyncio.wait(
        [
            asyncio.create_task(self.async_stop()),
            asyncio.create_task(self._task_manager.wait_for_exceptions())
        ],
        return_when=asyncio.FIRST_COMPLETED
    )

    if self._task_manager.exception_occured():
        for task, e, source_trace in self._task_manager.get_exceptions():
            print(source_trace)
            raise e
```

There are two places where exceptions may occur:

1. During the execution of the Net (in `net_IO_handler`)
2. After execution, as the net is stopping (in `self.async_stop`).

If an exception bubbles up via the chain of TaskManagers, the `asyncio.wait` stops and raises them.

- Nodes now have an `address` field that gives a full string address to the node, using its own and its parents ids. Might also define another address that returns the sequence of ids as a tuple also, instead of as a string.