feat: changed how ports are exposed
As before the `port.get` and the `port.put` methods of `Port` are presumed to be called from *within* the component process.

This is why we have this code in `Port.__init__`:

```python
if self.is_input_port:
    self.get = self._get
    self.get_and_consume = self._get_and_consume
else:
    self.put = self._put
    self.put_value = self._put_value
```

(eventually I should create a dedicated ComponentPort class, and move this logic to that)

Correspondingly, `NodePort` now has

```python
if not self.is_input_port:
    self.get = self._get_to_external
    self.get_and_consume = self._get_and_consume_to_external
else:
    self.put = self._put_from_external
    self.put_value = self._put_value_to_external
```

Which assumes that `get` and `put` are called from *outside* the Node logic.

Therefore, all internal sending of packets should be via the get and put methods prepended with `_` (so `_get` and `_put`). Specifically, `NodePort.put` and `NodePort.get` will always
assume that packets are coming from outside of the net, and will raise an exception if this is not the case.

I've also removed the `BaseComponent.put_packet` and `BaseComponent.get_packet` methods, as they are redundant.