feat(comp): Component factories now set the __module__ attribute of the components
Component factories will now by default set the __module__ attribute of the components to match the caller of `BaseComponent.create_component`.

That is `BaseComponent._create_component_class` will set `comp.__module__` to the *caller of the caller* of `BaseComponent._create_component_class` (presuming that `BaseComponent._create_component_class` is always called from `BaseComponent.create_component`).

This is a bit of a hack, but something like this is necessary to allow for components created from factories to be pickled properly.